refactor: move isConnectorTypeTrackable to lib/connector_providers
