+++ front/lib/connector_providers.ts
@@ -495,3 +495,27 @@ export function getDefaultDataSourceDescription(
     ? `Managed Data Source for ${provider} (${suffix})`
     : `Managed Data Source for ${provider}`;
 }
+
+export function isConnectorTypeTrackable(
+  connectorType: ConnectorProvider
+): boolean {
+  switch (connectorType) {
+    case "google_drive":
+    case "github":
+    case "notion":
+    case "microsoft":
+    case "confluence":
+    case "intercom":
+    case "webcrawler":
+    case "snowflake":
+    case "zendesk":
+    case "bigquery":
+    case "salesforce":
+    case "gong":
+      return true;
+    case "slack":
+      return false;
+    default:
+      assertNever(connectorType);
+  }
+}
